add delete checkboxes

+++ task1/src/app/admin-panel/admin-panel.component.ts
@@ -7,6 +7,7 @@ import { DialogDataAdd} from '../add-user-data/add-user-data.component';
 import { EditUserDataComponent } from '../edit-user-data/edit-user-data.component';
 import { DeleteUserDataComponent } from '../delete-user-data/delete-user-data.component';
 import { HeaderObserveService } from '../header-observe.service';
+import { SelectionModel } from '@angular/cdk/collections';
 
 export interface UsersElements {
   //*поменять типы
@@ -23,7 +24,6 @@ export interface DialogData {
 }
 
 const ELEMENT_DATA: UsersElements[] = [];
-
 @Component({
   selector: 'app-admin-panel',
   templateUrl: './admin-panel.component.html',
@@ -32,8 +32,9 @@ const ELEMENT_DATA: UsersElements[] = [];
 export class AdminPanelComponent implements OnInit {
   id : number;
   email: number;
-  displayedColumns: string[] = ['id', 'email', 'password', 'age', 'telephone', 'action'];
+  displayedColumns: string[] = ['select', 'id', 'email', 'password', 'age', 'telephone', 'action'];
   dataSource = new MatTableDataSource<UsersElements>(ELEMENT_DATA);
+  selection = new SelectionModel<UsersElements>(true, []);
   
   constructor(public dialog: MatDialog, private adminService: AdminToolsService, private infoService: HeaderObserveService) { }
 
@@ -76,7 +77,8 @@ export class AdminPanelComponent implements OnInit {
     const dialogRef = this.dialog.open(DeleteUserDataComponent, { data: { id: user.id, email: user.email }});
 
       dialogRef.afterClosed().subscribe(async result => {
-      if(result){
+        //duplication
+        if (result) {
         let response = await this.deleteUser(user);
         if (response){
           let data = this.dataSource.data;
@@ -88,12 +90,10 @@ export class AdminPanelComponent implements OnInit {
   }
 
   openDialogEdit(user){
-    debugger;
 
     //БАГ при вызове индекса второй раз прилетит -1;
     //Происходит из-за того что таблицы изменилась, а элемента_дата константа потому и не находит.
-    let index = ELEMENT_DATA.indexOf(user);
-    
+
     const dialogRef = this.dialog.open(EditUserDataComponent, 
       { data: 
         { 
@@ -114,30 +114,78 @@ export class AdminPanelComponent implements OnInit {
       fetch('http://localhost:3000/users').then(item => item.json()).then(elem => {
         //Заполняем массив полученными данными
         return elem.map(item => data.push(item))
-
       })
       .then(() => this.dataSource = new MatTableDataSource<UsersElements>(data))
         .then(() => this.dataSource.paginator = this.paginator)
         .then(() => {
-          dataImg = this.dataSource.filteredData;
-          for(let i in dataImg){
-            console.log(dataImg[i].img);
-            
+          //ИСПРАВИТЬ баги со сменой автарки 
+          if(user.email === "admin@gmail.com"){
+            let index = data.findIndex(i => i.email === user.email)
+            dataImg = this.dataSource.filteredData;
+            this.infoService.anounceHeaderImg(dataImg[index].img);
           }
-          console.log(index);
-          
         })
       })
   }
+
+   deleteRows() {
+     
+    this.selection.selected.forEach((user, i) => {
+
+setTimeout(() => {
   
+      let index = this.dataSource.data.indexOf(user);
+
+      this.deleteUser(user);
+       //duplication
+       
+          this.dataSource.data.splice(index, 1);
+          this.dataSource = new MatTableDataSource<UsersElements>(this.dataSource.data);
+          this.selection = new SelectionModel<UsersElements>(true, []);
+}, 100 * (i + 1));
+
+
+       
+     
+     
+  })
+}
   async deleteUser(user){
     //ИЗМЕНИТЬ
-    return fetch(`http://localhost:3000/users/${user.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json"
-      },
-      body: null
-    })
+    if(user.email != "admin@gmail.com"){
+      return fetch(`http://localhost:3000/users/${user.id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json"
+        }
+      })
+    }
+    else{
+      console.log("Удалить админа нельзя");
+      
+    }
+  }
+
+
+  /** Whether the number of selected elements matches the total number of rows. */
+  isAllSelected() {
+    const numSelected = this.selection.selected.length;
+    const numRows = this.dataSource.data.length;   
+    return numSelected === numRows;
+  }
+
+  /** Selects all rows if they are not all selected; otherwise clear selection. */
+  masterToggle() {
+    this.isAllSelected() ?
+      this.selection.clear() :
+      this.dataSource.data.forEach(row => this.selection.select(row));
+  }
+
+  /** The label for the checkbox on the passed row */
+  checkboxLabel(row?: UsersElements): string {
+    if (!row) {
+      return `${this.isAllSelected() ? 'select' : 'deselect'} all`;
+    }
+    return `${this.selection.isSelected(row) ? 'deselect' : 'select'} row ${row.id + 1}`;
   }
 }
